requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
